# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/study-planner": {
            "get": {
                "operationId": "StudyPlannerController_getTasks",
                "summary": "Get all tasks paginated",
                "parameters": [
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "example": 10,
                        "description": "Number of items per page",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "offset",
                        "required": true,
                        "in": "query",
                        "example": 1,
                        "description": "Page number",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all tasks paginated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. request parameters have something wrong"
                    }
                },
                "tags": [
                    "study-planner"
                ]
            },
            "post": {
                "operationId": "StudyPlannerController_createTask",
                "summary": "Create a new task",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddTaskDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "a task has been successfully created."
                    },
                    "400": {
                        "description": "Bad request. request parameters have something wrong"
                    }
                },
                "tags": [
                    "study-planner"
                ]
            }
        },
        "/api/study-planner/{task_id}": {
            "get": {
                "operationId": "StudyPlannerController_getTaskById",
                "summary": "Get a specific task",
                "parameters": [
                    {
                        "name": "task_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully get a specific task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "examples": {
                                        "application/json": {
                                            "status": "success",
                                            "data": {
                                                "task_id": 1,
                                                "title": "task 1",
                                                "description": "task 1 description",
                                                "start_date": "2021-05-19T00:00:00.000Z",
                                                "due_date": "2021-05-20T00:00:00.000Z",
                                                "status": "in-progress",
                                                "user_id": 1
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. request parameters have something wrong"
                    }
                },
                "tags": [
                    "study-planner"
                ]
            },
            "patch": {
                "operationId": "StudyPlannerController_updateTask",
                "summary": "Update a task",
                "parameters": [
                    {
                        "name": "task_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task has been successfully updated"
                    },
                    "400": {
                        "description": "Bad request. request parameters have something wrong"
                    }
                },
                "tags": [
                    "study-planner"
                ]
            },
            "delete": {
                "operationId": "StudyPlannerController_deleteTask",
                "summary": "Delete a task",
                "parameters": [
                    {
                        "name": "task_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task has been successfully deleted"
                    },
                    "400": {
                        "description": "Bad request. request parameters have something wrong"
                    }
                },
                "tags": [
                    "study-planner"
                ]
            }
        }
    },
    "info": {
        "title": "INTELLI GP",
        "description": "The INTELLI-GP API documentation",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "AddTaskDto": {
                "type": "object",
                "properties": {
                    "Title": {
                        "type": "string",
                        "description": "The title of the task",
                        "maxLength": 255,
                        "example": "task 1"
                    },
                    "Description": {
                        "type": "string",
                        "description": "The description of the task",
                        "maxLength": 255,
                        "example": "task 1 for CS256 assignment"
                    },
                    "StartDate": {
                        "type": "string",
                        "description": "The start date of the task in the format yyyy-mm-ddThh:mm",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}$",
                        "example": "2021-01-01T00:00"
                    },
                    "DueDate": {
                        "type": "string",
                        "description": "The due date of the task in the format yyyy-mm-ddThh:mm",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}$",
                        "example": "2021-01-01T00:00"
                    },
                    "Status": {
                        "type": "string",
                        "description": "The status of the task",
                        "maxLength": 255,
                        "example": "in progress"
                    }
                },
                "required": [
                    "Title",
                    "StartDate",
                    "DueDate",
                    "Status"
                ]
            },
            "UpdateTaskDto": {
                "type": "object",
                "properties": {
                    "Title": {
                        "type": "string",
                        "description": "The title of the task",
                        "maxLength": 255,
                        "example": "task 1"
                    },
                    "Description": {
                        "type": "string",
                        "description": "The description of the task",
                        "maxLength": 255,
                        "example": "task 1 for CS256 assignment"
                    },
                    "StartDate": {
                        "type": "string",
                        "description": "The start date of the task in the format yyyy-mm-ddThh:mm",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}$",
                        "example": "2021-01-01T00:00"
                    },
                    "DueDate": {
                        "type": "string",
                        "description": "The due date of the task in the format yyyy-mm-ddThh:mm",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}$",
                        "example": "2021-01-01T00:00"
                    },
                    "Status": {
                        "type": "string",
                        "description": "The status of the task",
                        "maxLength": 255,
                        "example": "in progress"
                    }
                }
            }
        }
    }
}